Quantiles class rework
Changed floats to doubles everywhere possible. Removed casts where possible.

+++ src/BasicDataPlot.cpp
@@ -139,19 +139,19 @@ void BasicDataPlot::setNewData(QVector<QPointF> data,
 
     if (data.size() > 0)
     {
-        float min = quantiles.minX_;
-        float max = quantiles.maxX_;
+        double min = quantiles.minX_;
+        double max = quantiles.maxX_;
 
-        qVector.append(QPointF(static_cast<double>(min), static_cast<double>(quantiles.q50_)));
-        qVector.append(QPointF(static_cast<double>(max), static_cast<double>(quantiles.q50_)));
+        qVector.append(QPointF(min, quantiles.q50_));
+        qVector.append(QPointF(max, quantiles.q50_));
         plotQ50_.setSamples(qVector);
         qVector.clear();
-        qVector.append(QPointF(static_cast<double>(min), static_cast<double>(quantiles.q25_)));
-        qVector.append(QPointF(static_cast<double>(max), static_cast<double>(quantiles.q25_)));
+        qVector.append(QPointF(min, quantiles.q25_));
+        qVector.append(QPointF(max, quantiles.q25_));
         plotQ25_.setSamples(qVector);
         qVector.clear();
-        qVector.append(QPointF(static_cast<double>(min), static_cast<double>(quantiles.q75_)));
-        qVector.append(QPointF(static_cast<double>(max), static_cast<double>(quantiles.q75_)));
+        qVector.append(QPointF(min, quantiles.q75_));
+        qVector.append(QPointF(max, quantiles.q75_));
         plotQ75_.setSamples(qVector);
 
         plotLinearRegression_.setSamples(linearRegression);
